Replaces the editor textarea with Lexical (contenteditable)
This is mostly a 1:1 replacement, just to setup future work. You can see the
result of this by entering a Heading with "# Heading". One interesting change is
that the editor is now "uncontrolled" - only the `onChange` handler needs to be
added, not the `value`.

Lexical is very heavyweight but *sigh* is needed for contenteditable. Yay web!