Assistant keeps history over multiple messages
I was inspired by re-reading ["The Art of the Propagator"](https://dspace.mit.edu/handle/1721.1/44215) to have some kind of "merge" function for the final value so that it would eventually "converge" on a correct value. After humming and hawing over it for a few days I realized that the best way of doing this is to just add to context. The model is "merging" the latest input and refining the output. Seems obvious in retrospect and I'm glad I didn't go down the path of having another LLM call to "merge two outputs". The whole tool now feels more magical in a way that doesn't change the interface at all, which in my opinion is a HUGE win.

I added a little shortcut to empty context when the input resets back to "". I'm not 100% sure I like it long term because it's not particularly discoverable, but it feels _fast_ for the moment.

(Also a few other misc. changes to Portals but they'll get some more work in a bit.)