Wire up simple demo with a hard-coded Portal
The main goal of this PR is to start to get this into a demo-able form so I can get early feedback on the idea from other people.

This passes a value from the Portal up through the editor, which then invalidates the propagator. It's a bit buggy at the moment because the document changes faster than the Assistant's debounce so the propagator calculation gets continually cancelled.

There are some funky edge cases with loading states but I think I can work on those later.