Futz with the throttling and debounce to make the clock demo mostly work
The first change, to only call `onUpdate` when the editor's markdown value has actually changed, is good. This could be done on the caller's side too but it makes sense to do it as early as possible to prevent re-renders everywhere.

The second change, updating throttling and debouncing really doesn't work well. It's very dependent on the speed at which folks type, the speed at which portal values can update and the speed at which the propagator can calculate it's new value. The best way to handle this IMHO is to have some kind of a "lock" (or queue) that prevents the propagator from updating the portal value until the previous update has been processed. That would decouple saturating the propagator from any UI updates from the editor. TBD.