End-to-end binary example
This doesn't work well but is meant to show the end-to-end process of using Monument. Basically, point it at a directory of Markdown files and it starts to process them, feeding them to an LLM. It's nice because then you can reference the results from a new document - the output is just files on disk. Definitely mixing and matching here, but isn't that what experiments are all about? Breaking down barriers between source and build outputs.

Some things it does poorly:

1. No "document sharing". Each source document maintains it's own graph.
2. The CLI output is gross because it's just spawning subprocesses and piping them together. It's one step above a shell script!
3. There's concurrency issues with the model output.